pyproject.toml
[tool.poetry]
name = "telephonist"
version = "0.1.0"
description = ""
authors = ["MaratBR <marat01q@gmail.com>"]
license = "Apache-2.0 License"

[tool.poetry.dependencies]
python = "^3.8"
beanie = "^1.7.0"
fastapi = "^0.70.0"
pydantic = "^1.8.2"
uvicorn = "^0.16.0"
email-validator = "1.1.3"
passlib = "1.7.4"
aioredis = "2.0.0"
nanoid = "^2.0.0"
python-multipart = "^0.0.5"
broadcaster = "^0.2.0"
websockets = "^10.0"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
pyhumps = "^3.0.2"
docker = "^5.0.3"
pybranca = "^0.5.0"
paseto = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
docker = "^5.0.3"
pytest-asyncio = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
